annadido IRecordRepository y RecordRepository generico
