+++ src/SchoolManagementSystem.Infrastructure/Repositories/Entities/CourseRepository.cs
@@ -3,11 +3,11 @@ using SchoolManagementSystem.Domain.Interfaces;
 using SchoolManagementSystem.Infrastructure.Data;
 using SchoolManagementSystem.Application.Repositories_Interfaces;
 
-namespace SchoolManagementSystem.Infrastructure.Repositories;
+namespace SchoolManagementSystem.Infrastructure.Repositories.Entities;
 
-public class TeacherRepository : CrudRepository<Teacher>, ITeacherRepository
+public class CourseRepository : CrudRepository<Course>, ICourseRepository
 {
-    public TeacherRepository(IObjectContext context) : base(context)
+    public CourseRepository(IObjectContext context) : base(context)
     {
 
     }
