+++ src/SchoolManagementSystem.Infrastructure/Repositories/Entities/StudentRepository.cs
@@ -3,7 +3,7 @@ using SchoolManagementSystem.Domain.Interfaces;
 using SchoolManagementSystem.Infrastructure.Data;
 using SchoolManagementSystem.Application.Repositories_Interfaces;
 
-namespace SchoolManagementSystem.Infrastructure.Repositories;
+namespace SchoolManagementSystem.Infrastructure.Repositories.Entities;
 
 public class StudentRepository : CrudRepository<Student>, IStudentRepository
 {
