+++ src/SchoolManagementSystem.Infrastructure/Repositories/Entities/TuitorRepository.cs
@@ -3,11 +3,11 @@ using SchoolManagementSystem.Domain.Interfaces;
 using SchoolManagementSystem.Infrastructure.Data;
 using SchoolManagementSystem.Application.Repositories_Interfaces;
 
-namespace SchoolManagementSystem.Infrastructure.Repositories;
+namespace SchoolManagementSystem.Infrastructure.Repositories.Entities;
 
-public class PositionRepository : CrudRepository<Position>, IPositionRepository
+public class TuitorRepository : CrudRepository<Tuitor>, ITuitorRepository
 {
-    public PositionRepository(IObjectContext context) : base(context)
+    public TuitorRepository(IObjectContext context) : base(context)
     {
 
     }
